huggingface-convenient-usage/DepthAnythingV2/requirements.txt
# 如果出现 KeyError: "Unknown task depth-estimation, ..."错误，尝试重新安装 transformers（卸了再重装）
transformers
pillow
torch
torchvision

# 如果需要在 jupyter notebook 中演示，还需要安装以下包：
jupyter
matplotlib